Canonicalize by ::realpath on server side
Client cannot call ::realpath because file scanning is not working with
client permission but server permission.

Client just make path to absolute (prepend current directory if relative
path comes in to parameter) and give them to server to get canonicalized
file set.

It's same to files_async and dirs_async but we need to erase
subdirectory in dirs_async case after canonicalization done.

Change-Id: I15ddd7e8816ee068c9e4439eea51bbe8c0a2602b
Signed-off-by: Kyungwook Tak <k.tak@samsung.com>